requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python
numpy
tqdm
pycocotools
pyyaml
pillow
twilio
scipy
filterpy>=1.4.5
#pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118